Move CTCommons project into ClassTranscribeDatabase project
