--- CTCommons/CTCommons.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <!-- Needed for Visual Studio for Mac to recognize the macOS project configurations. -->
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' " />
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' " />
-  
-  <ItemGroup>
-    <PackageReference Include="Box.V2.Core" Version="3.22.0" />
-    <PackageReference Include="Google.Protobuf" Version="3.17.3" />
-    <PackageReference Include="Grpc.Core" Version="2.39.1" />
-    <PackageReference Include="Grpc.Tools" Version="2.39.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.CognitiveServices.Speech" Version="1.12.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ClassTranscribeDatabase\ClassTranscribeDatabase.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <Protobuf Include="..\ct.proto" GrpcServices="Client" />
-  </ItemGroup>
-
-</Project>
