+++ ClassTranscribeDatabase/Services/BoxAPI.cs
@@ -1,7 +1,6 @@
 ﻿using Box.V2;
 using Box.V2.Auth;
 using Box.V2.Config;
-using ClassTranscribeDatabase;
 using ClassTranscribeDatabase.Models;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -9,7 +8,7 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace CTCommons
+namespace ClassTranscribeDatabase.Services
 {
     public class BoxAPI
     {
