+++ ClassTranscribeServer/ClassTranscribeServer.csproj
@@ -77,7 +77,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ClassTranscribeDatabase\ClassTranscribeDatabase.csproj" />
-    <ProjectReference Include="..\CTCommons\CTCommons.csproj" />
   </ItemGroup>
 
   <ItemGroup>
