+++ ClassTranscribeServer/Utils/WakeDownloader.cs
@@ -1,4 +1,4 @@
-﻿using CTCommons;
+﻿using ClassTranscribeDatabase.Services;
 using Newtonsoft.Json.Linq;
 using static ClassTranscribeDatabase.CommonUtils;
