+++ DevExperiments/DevExperiments.csproj
@@ -20,7 +20,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ClassTranscribeDatabase\ClassTranscribeDatabase.csproj" />
-    <ProjectReference Include="..\CTCommons\CTCommons.csproj" />
   </ItemGroup>
 
   <ItemGroup Condition="Exists('..\vs_appsettings.json')">
