+++ TaskEngine/TaskEngine.csproj
@@ -30,7 +30,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ClassTranscribeDatabase\ClassTranscribeDatabase.csproj" />
-    <ProjectReference Include="..\CTCommons\CTCommons.csproj" />
   </ItemGroup>
 
   <ItemGroup Condition="Exists('..\vs_appsettings.json')">
