+++ TaskEngine/Tasks/BuildElasticIndexTask.cs
@@ -1,15 +1,15 @@
 ﻿using ClassTranscribeDatabase;
-using CTCommons;
+using ClassTranscribeDatabase.Services;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Configuration;
-using System.Threading.Tasks;
-using static ClassTranscribeDatabase.CommonUtils;
-using System.Linq;
+using Microsoft.Extensions.Logging;
 using Nest;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Threading.Tasks;
+using static ClassTranscribeDatabase.CommonUtils;
 
 namespace TaskEngine.Tasks
 {
