+++ TaskEngine/Tasks/DownloadPlaylistInfoTask.cs
@@ -1,19 +1,18 @@
 ﻿using Box.V2.Models;
 using ClassTranscribeDatabase;
 using ClassTranscribeDatabase.Models;
+using ClassTranscribeDatabase.Services;
 using Grpc.Core;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
-using CTCommons.Grpc;
-using CTCommons;
 using static ClassTranscribeDatabase.CommonUtils;
-using System.Diagnostics.CodeAnalysis;
 
 
 namespace TaskEngine.Tasks
