+++ TaskEngine/Tasks/ExampleTask.cs
@@ -1,11 +1,11 @@
 ﻿using ClassTranscribeDatabase;
-using CTCommons;
+using ClassTranscribeDatabase.Services;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using System.Threading.Tasks;
-using static ClassTranscribeDatabase.CommonUtils;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Threading.Tasks;
+using static ClassTranscribeDatabase.CommonUtils;
 
 namespace TaskEngine.Tasks
 {
