+++ TaskEngine/Tasks/GenerateVTTFileTask.cs
@@ -1,11 +1,11 @@
 ﻿using ClassTranscribeDatabase;
 using ClassTranscribeDatabase.Models;
-using CTCommons;
+using ClassTranscribeDatabase.Services;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using static ClassTranscribeDatabase.CommonUtils;
-using System.Diagnostics.CodeAnalysis;
 
 
 namespace TaskEngine.Tasks
