+++ TaskEngine/Tasks/RabbitMQTask.cs
@@ -1,10 +1,10 @@
-﻿using CTCommons;
+﻿using ClassTranscribeDatabase.Services;
 using Microsoft.Extensions.Logging;
 using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using static ClassTranscribeDatabase.CommonUtils;
-using System.Diagnostics.CodeAnalysis;
-using System.Collections.Generic;
 
 namespace TaskEngine
 {
