+++ TaskEngine/Tasks/SceneDetectionTask.cs
@@ -1,14 +1,13 @@
-﻿using System.Collections.Generic;
-using System.IO;
-using ClassTranscribeDatabase;
+﻿using ClassTranscribeDatabase;
 using ClassTranscribeDatabase.Models;
+using ClassTranscribeDatabase.Services;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
-using System.Threading.Tasks;
-using CTCommons.Grpc;
-using static ClassTranscribeDatabase.CommonUtils;
-using CTCommons;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
+using System.IO;
+using System.Threading.Tasks;
+using static ClassTranscribeDatabase.CommonUtils;
 
 
 namespace TaskEngine.Tasks
